+++ src/JMcDM.jl
@@ -168,7 +168,7 @@ import .SCDM: SavageResult, HurwiczResult, MLEResult, ExpectedRegretResult
 import .SCDM:
     laplace, maximax, maximin, minimax, minimin, savage, hurwicz, mle, expectedregret
 
-import .GreyNumbers: GreyNumber, kernel, whitenizate
+import .GreyNumbers: GreyNumber, kernel, whitenizate, simplify
 
 import .Utilities: *
 using .Utilities
@@ -316,7 +316,7 @@ import JMcDM.Copeland: copeland
 export copeland
 
 # export Grey Number elements
-export GreyNumber, kernel, whitenizate
+export GreyNumber, kernel, whitenizate, simplify
 
 export Normalizations